+++ src/pyochain/__init__.py
@@ -5,8 +5,9 @@ from ._dict import Dict
 from ._iter import Iter
 from ._range import Range
 from ._seq import Seq
-from ._set import PyoItemsView, PyoKeysView, PyoValuesView, Set, SetMut
+from ._set import Set, SetMut
 from ._vec import Vec
+from ._views import PyoItemsView, PyoKeysView, PyoValuesView
 from .rs import (
     NONE,
     Err,